+++ src/main/java/io/seqware/pancancer/DEWrapperWorkflow.java
@@ -77,8 +77,6 @@ public class DEWrapperWorkflow extends AbstractWorkflowDataModel {
     private String s3Key = null;
     private String s3SecretKey = null;
     private String uploadS3BucketPath = null;
-    // workflows to run
-    private Boolean runDkfz = true;
     // docker names, do not specify defaults here. They are misleading and
     // they will be overridden by the embedded default ini anyways
     private String dkfzDockerName;
@@ -115,9 +113,9 @@ public class DEWrapperWorkflow extends AbstractWorkflowDataModel {
             // these variables are those extra required for EMBL upload
             this.uploadServer = getProperty("uploadServer");
             StringBuilder metadataURLBuilder = new StringBuilder();
-            metadataURLBuilder.append(uploadServer).append("/cghub/metadata/analysisFull/").append(controlAnalysisId);
+            metadataURLBuilder.append(gnosServer).append("/cghub/metadata/analysisFull/").append(controlAnalysisId);
             for (String id : Lists.newArrayList(getProperty("tumourAnalysisIds").split(","))) {
-                metadataURLBuilder.append(",").append(uploadServer).append("/cghub/metadata/analysisFull/").append(id);
+                metadataURLBuilder.append(",").append(gnosServer).append("/cghub/metadata/analysisFull/").append(id);
             }
             this.metadataURLs = metadataURLBuilder.toString();
             this.tumorAliquotIds = Lists.newArrayList(getProperty("tumourAliquotIds").split(","));
@@ -233,21 +231,16 @@ public class DEWrapperWorkflow extends AbstractWorkflowDataModel {
         Job emblJob = runEMBLWorkflow(lastDownloadDataJob);
         Job dkfzJob = null;
 
-        Job uploadEMBLJob = this.uploadEMBLJob();
-        uploadEMBLJob.addParent(emblJob);
+        // call the DKFZ workflow
+        dkfzJob = runDKFZWorkflow(emblJob);
 
-        Job uploadDKFZJob = null;
-        if (runDkfz) {
-            // call the DKFZ workflow
-            dkfzJob = runDKFZWorkflow(emblJob);
-            uploadDKFZJob = this.uploadDKFZJob();
-            uploadDKFZJob.addParent(emblJob);
-            uploadDKFZJob.addParent(dkfzJob);
-            uploadEMBLJob.addParent(dkfzJob);
-        }
+        // common upload job
+        Job uploadJob = uploadJob();
+        uploadJob.addParent(emblJob);
+        uploadJob.addParent(dkfzJob);
 
         // now cleanupJob
-        cleanupWorkflow(uploadEMBLJob, uploadDKFZJob);
+        cleanupWorkflow(uploadJob);
 
     }
 
@@ -328,6 +321,157 @@ public class DEWrapperWorkflow extends AbstractWorkflowDataModel {
 
         emblJob.addParent(previousJobPointer);
         return emblJob;
+    }
+
+    private Job uploadJob() throws RuntimeException {
+
+        // upload the EMBL and DKFZ results together
+
+        List<String> vcfs = new ArrayList<>();
+        List<String> tbis = new ArrayList<>();
+        List<String> tars = new ArrayList<>();
+        List<String> vcfmd5s = new ArrayList<>();
+        List<String> tbimd5s = new ArrayList<>();
+        List<String> tarmd5s = new ArrayList<>();
+
+        // FIXME: really just need one timing file not broken down by tumorAliquotID! This will be key for multi-tumor donors
+        String qcJson = null;
+        String qcJsonSingle = null;
+        String timingJson = null;
+
+        // FIXME: these don't quite follow the naming convention
+        for (String tumorAliquotId : tumorAliquotIds) {
+
+            // DELLY FILES
+            // String baseFile = "/workflow_data/" + tumorAliquotId + ".embl-delly_1-0-0-preFilter."+formattedDate;
+            String baseFile = tumorAliquotId + ".embl-delly_1-3-0-preFilter." + formattedDate;
+
+            qcJson = "`find . | grep " + baseFile + ".sv.qc.json | head -1`";
+            timingJson = "`find . | grep " + baseFile + ".sv.timing.json | head -1`";
+
+            vcfs.add(baseFile + ".germline.sv.vcf.gz");
+            vcfs.add(baseFile + ".sv.vcf.gz");
+            vcfs.add(baseFile + ".somatic.sv.vcf.gz");
+
+            vcfmd5s.add(baseFile + ".germline.sv.vcf.gz.md5");
+            vcfmd5s.add(baseFile + ".sv.vcf.gz.md5");
+            vcfmd5s.add(baseFile + ".somatic.sv.vcf.gz.md5");
+
+            tbis.add(baseFile + ".germline.sv.vcf.gz.tbi");
+            tbis.add(baseFile + ".sv.vcf.gz.tbi");
+            tbis.add(baseFile + ".somatic.sv.vcf.gz.tbi");
+
+            tbimd5s.add(baseFile + ".germline.sv.vcf.gz.tbi.md5");
+            tbimd5s.add(baseFile + ".sv.vcf.gz.tbi.md5");
+            tbimd5s.add(baseFile + ".somatic.sv.vcf.gz.tbi.md5");
+
+            tars.add(baseFile + ".germline.sv.readname.txt.tar.gz");
+            tarmd5s.add(baseFile + ".germline.sv.readname.txt.tar.gz.md5");
+
+            tars.add(baseFile + ".germline.sv.bedpe.txt.tar.gz");
+            tarmd5s.add(baseFile + ".germline.sv.bedpe.txt.tar.gz.md5");
+
+            tars.add(baseFile + ".somatic.sv.readname.txt.tar.gz");
+            tarmd5s.add(baseFile + ".somatic.sv.readname.txt.tar.gz.md5");
+
+            tars.add(baseFile + ".somatic.sv.bedpe.txt.tar.gz");
+            tarmd5s.add(baseFile + ".somatic.sv.bedpe.txt.tar.gz.md5");
+
+            tars.add(baseFile + ".sv.cov.plots.tar.gz");
+            tarmd5s.add(baseFile + ".sv.cov.plots.tar.gz.md5");
+
+            tars.add(baseFile + ".sv.cov.tar.gz");
+            tarmd5s.add(baseFile + ".sv.cov.tar.gz.md5");
+
+
+            // DKFZ FILES
+            // String baseFile = "/workflow_data/" + tumorAliquotId + ".dkfz-";
+            baseFile = tumorAliquotId + ".dkfz-";
+
+            qcJson = tumorAliquotId + ".qc_metrics.dkfz.json";
+            qcJsonSingle = tumorAliquotId + ".qc_metrics.dkfz.single.json";
+            timingJson = "timing.json";
+
+            // VCF
+            vcfs.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".germline.indel.vcf.gz");
+            vcfs.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.vcf.gz");
+            vcfs.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".germline.snv_mnv.vcf.gz");
+            vcfs.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.snv_mnv.vcf.gz");
+            vcfs.add(baseFile + "copyNumberEstimation_" + DKFZ_VERSION + "." + formattedDate + ".somatic.cnv.vcf.gz");
+
+            // VCF MD5
+            vcfmd5s.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".germline.indel.vcf.gz.md5");
+            vcfmd5s.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.vcf.gz.md5");
+            vcfmd5s.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".germline.snv_mnv.vcf.gz.md5");
+            vcfmd5s.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.snv_mnv.vcf.gz.md5");
+            vcfmd5s.add(baseFile + "copyNumberEstimation_" + DKFZ_VERSION + "." + formattedDate + ".somatic.cnv.vcf.gz.md5");
+
+            // Tabix
+            tbis.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.vcf.gz.tbi");
+            tbis.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.vcf.gz.tbi");
+            tbis.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".germline.snv_mnv.vcf.gz.tbi");
+            tbis.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.snv_mnv.vcf.gz.tbi");
+            tbis.add(baseFile + "copyNumberEstimation_" + DKFZ_VERSION + "." + formattedDate + ".somatic.cnv.vcf.gz.tbi");
+
+            // Tabix MD5
+            tbimd5s.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.vcf.gz.tbi.md5");
+            tbimd5s.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.vcf.gz.tbi.md5");
+            tbimd5s.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".germline.snv_mnv.vcf.gz.tbi.md5");
+            tbimd5s.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.snv_mnv.vcf.gz.tbi.md5");
+            tbimd5s.add(baseFile + "copyNumberEstimation_" + DKFZ_VERSION + "." + formattedDate + ".somatic.cnv.vcf.gz.tbi.md5");
+
+            // Tarballs
+            tars.add(baseFile + "copyNumberEstimation_" + DKFZ_VERSION + "." + formattedDate + ".somatic.cnv.tar.gz");
+            tars.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.tar.gz");
+            tars.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.snv_mnv.tar.gz");
+
+            // Tarballs MD5
+            tarmd5s.add(baseFile + "copyNumberEstimation_" + DKFZ_VERSION + "." + formattedDate + ".somatic.cnv.tar.gz.md5");
+            tarmd5s.add(baseFile + "indelCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.indel.tar.gz.md5");
+            tarmd5s.add(baseFile + "snvCalling_" + DKFZ_VERSION + "." + formattedDate + ".somatic.snv_mnv.tar.gz.md5");
+
+        }
+        // perform upload to GNOS
+        // FIXME: hardcoded versions, URLs, etc
+        Job uploadJob = this.getWorkflow().createBashJob("upload");
+        // params
+        StringBuilder overrideTxt = new StringBuilder();
+        if (this.studyRefnameOverride != null) {
+            overrideTxt.append(" --study-refname-override ").append(this.studyRefnameOverride);
+        }
+        if (this.analysisCenterOverride != null) {
+            overrideTxt.append(" --analysis-center-override ").append(this.analysisCenterOverride);
+        }
+        // Now do the upload based on the destination chosen
+        // NOTE: I'm using the wrapper workflow version here so it's immediately obvious what wrapper was used
+        if (LOCAL.equalsIgnoreCase(uploadDestination)) {
+
+            // using hard links so it spans multiple exported filesystems to Docker
+            uploadJob = utils.localUploadJob(uploadJob, SHARED_WORKSPACE_ABSOLUTE, pemFile, metadataURLs, vcfs, vcfmd5s, tbis, tbimd5s,
+                    tars, tarmd5s, uploadServer, Version.SEQWARE_VERSION, vmInstanceType, vmLocationCode, overrideTxt.toString(),
+                    UPLOAD_ARCHIVE_IN_CONTAINER, gnosTimeoutMin, gnosRetries, qcJson, timingJson, Version.WORKFLOW_SRC_URL,
+                    Version.WORKFLOW_URL, Version.WORKFLOW_NAME, Version.WORKFLOW_VERSION, gnosDownloadName,
+                    this.localXMLMetadataPath, this.localXMLMetadataFiles);
+
+        } else if (GNOS.equalsIgnoreCase(uploadDestination)) {
+
+            uploadJob = utils.gnosUploadJob(uploadJob, SHARED_WORKSPACE_ABSOLUTE, pemFile, metadataURLs, vcfs, vcfmd5s, tbis, tbimd5s,
+                    tars, tarmd5s, uploadServer, Version.SEQWARE_VERSION, vmInstanceType, vmLocationCode, overrideTxt.toString(),
+                    gnosTimeoutMin, gnosRetries, qcJson, timingJson, Version.WORKFLOW_SRC_URL,
+                    Version.WORKFLOW_URL, Version.WORKFLOW_NAME, Version.WORKFLOW_VERSION, gnosDownloadName);
+
+        } else if (S3.equalsIgnoreCase(uploadDestination)) {
+
+            uploadJob = utils.s3UploadJob(uploadJob, SHARED_WORKSPACE_ABSOLUTE, pemFile, metadataURLs, vcfs, vcfmd5s, tbis, tbimd5s, tars,
+                    tarmd5s, uploadServer, Version.SEQWARE_VERSION, vmInstanceType, vmLocationCode, overrideTxt.toString(),
+                    UPLOAD_ARCHIVE_IN_CONTAINER, s3Key, s3SecretKey, uploadS3BucketPath, gnosTimeoutMin, gnosRetries, qcJson, timingJson,
+                    Version.WORKFLOW_SRC_URL, Version.WORKFLOW_URL, Version.WORKFLOW_NAME,  Version.WORKFLOW_VERSION,
+                    gnosDownloadName);
+
+        } else {
+            throw new RuntimeException("Don't know what upload type '" + uploadDestination + "' is!");
+        }
+        return uploadJob;
     }
 
     private Job uploadEMBLJob() throws RuntimeException {